+++ poc/CHANGES.md
@@ -1,4 +1,9 @@
 # Changes in the FPL Parser and Interpreter
+## 1.8.4
+* Refactoring / Feature:
+  * New class AuxEvaluation added
+  * Some simple evaluations on the predicate level added.
+  * New errors FplAxiomNotSatisfiable and FplTypeMismatch added.
 ## 1.8.3
 * Refactoring:
   * Introducing the 'evaluate' concept - a method to be implemented by (almost) each of the members of the symbol table that will implement the full semantics of each FPL language construct.